refactor(models): single shared ORM in shared/models, one image-type enum (Task 1.2)
Move the SQLAlchemy ORM (POI + subtypes, category, image, user, attribute,
primary_type, relationships) into shared/models/, defined once against one
shared declarative Base. Both backends re-export that Base from app.database
and import the classes through thin re-export shims at the old module paths, so
every existing import site keeps working. This removes the copy-pasted duplicate
ORMs that caused the sponsor_logo enum drift and the m_payphone rename outage.

- shared/models/base.py holds the single Base; admin Alembic env.py targets it
  unchanged (still Base.metadata via the poi shim).
- The pgvector embedding column stays UNMAPPED (raw-SQL read/write paths), with
  a comment explaining why; DDL-neutral (autogenerate diff identical to before).
- The two image-type enums collapse to one: admin schemas/image.py ImageTypeEnum
  now aliases shared ImageType (str-mixed for the Pydantic layer).
- App-only public form tables move to an app-local FormsBase so they do not
  collide on the shared Base metadata during the test harness module swap.
- gen_poi_registry.py reflects over shared/models/poi.py; registry output is
  byte-identical (zero diff).

No schema change; no migration.
