+++ nearby-admin/backend/app/models/user.py
@@ -1,16 +1,4 @@
-import uuid
-from sqlalchemy import Column, String, TIMESTAMP
-from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy.sql import func
+"""Re-export shim. The User ORM lives once in shared/models/user.py
+(Task 1.2). Edit shared/models/, not here."""
 
-from app.database import Base
-
-
-class User(Base):
-    __tablename__ = "users"
-
-    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    email = Column(String, unique=True, nullable=False, index=True)
-    hashed_password = Column(String, nullable=False)
-    role = Column(String, default='admin')  # 'admin', 'editor', etc.
-    created_at = Column(TIMESTAMP(timezone=True), server_default=func.now()) 
+from shared.models.user import User  # noqa: F401
