+++ shared/models/enums.py
@@ -28,7 +28,11 @@ class EventStatus(enum.Enum):
     UNOFFICIAL_PROPOSED = "Unofficial Proposed Date"
 
 
-class ImageType(enum.Enum):
+class ImageType(str, enum.Enum):
+    # ``str`` mixin: the admin Pydantic layer aliases this as ``ImageTypeEnum``
+    # (schemas/image.py) for request/response validation, which relied on
+    # str-enum semantics. SQLAlchemy still maps ``Enum(ImageType)`` by member
+    # name, so the DB ``imagetype`` labels are unchanged.
     main = "main"
     gallery = "gallery"
     entry = "entry"
