+++ shared/models/user.py
@@ -1,13 +1,16 @@
-# app/models/waitlist.py
 import uuid
 from sqlalchemy import Column, String, TIMESTAMP
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.sql import func
-from ..database import FormsBase as Base
 
-class WaitlistEntry(Base):
-    __tablename__ = "waitlist"
+from shared.models.base import Base
+
+
+class User(Base):
+    __tablename__ = "users"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     email = Column(String, unique=True, nullable=False, index=True)
+    hashed_password = Column(String, nullable=False)
+    role = Column(String, default='admin')  # 'admin', 'editor', etc.
     created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
